todo_backend/requirements.txt
Django==5.1.6
djangorestframework==3.15.0
gunicorn==23.0.0